+++ packages/python/src/components/NewExpression.tsx
@@ -8,9 +8,13 @@ import {
   FunctionCallExpression,
   FunctionCallExpressionProps,
 } from "./FunctionCallExpression.jsx";
-// TODO: Implement support for named parameters
+
 export interface NewExpressionProps extends FunctionCallExpressionProps {}
 
+// NewExpression is used to create new instances of classes in Python.
+// It is similar to FunctionCallExpression but specifically for class instantiation.
+// Args is a list arguments that can be either Values, which will render as positional arguments,
+// or VariableDeclarations, which will render as named arguments in the call statement.
 export function NewExpression(props: NewExpressionProps) {
   const sym = new PythonOutputSymbol("", {
     flags: OutputSymbolFlags.Transient,
