+++ packages/python/test/functioncallexpressions.test.tsx
@@ -2,6 +2,7 @@ import { d } from "@alloy-js/core/testing";
 import { describe, expect, it } from "vitest";
 import * as py from "../src/components/index.js";
 import { toSourceText } from "./utils.jsx";
+import { refkey } from "@alloy-js/core";
 
 describe("FunctionCallExpression", () => {
   it("renders", () => {
@@ -17,5 +18,115 @@ describe("FunctionCallExpression", () => {
     expect(result).toRenderTo(d`
       foo(a, b)
     `);
+  });
+
+  it("function call with variables", () => {
+    // Creating the reference separately so the naming policy doesn't interfere
+    const methodRef = refkey();
+    const result = toSourceText(
+      <py.StatementList>
+        <py.FunctionDeclaration name="runFunc" refkey={methodRef} />
+        <hbr />
+        <py.FunctionCallExpression
+          target={methodRef}
+          args={[
+            <py.Value jsValue={"A name"} />,
+            <py.Value jsValue={42} />,
+            <py.Value jsValue={true} />,
+          ]}
+        />
+      </py.StatementList>,
+    );
+    const expected = d`
+      def run_func():
+        pass
+
+
+      
+      run_func("A name", 42, True)
+    `;
+    expect(result).toRenderTo(expected);
+  });
+
+  it("function call with variables and assignment", () => {
+    // Creating the reference separately so the naming policy doesn't interfere
+    const methodRef = refkey();
+    const result = toSourceText(
+      <py.StatementList>
+        <py.FunctionDeclaration
+          name="runFunc"
+          returnType="str"
+          refkey={methodRef}
+          parameters={[
+            { name: "name", type: "str" },
+            { name: "number", type: "int" },
+            { name: "flag", type: "bool" },
+          ]}
+        />
+        <hbr />
+        <py.VariableDeclaration
+          name="result"
+          type={<py.Reference refkey={methodRef} />}
+          initializer={
+            <py.FunctionCallExpression
+              target={methodRef}
+              args={[
+                <py.Value jsValue={"A name"} />,
+                <py.Value jsValue={42} />,
+                <py.Value jsValue={true} />,
+              ]}
+            />
+          }
+        />
+      </py.StatementList>,
+    );
+    // TODO: Fix type once we handle types properly
+    const expected = d`
+      def run_func(name: str, number: int, flag: bool) -> str:
+        pass
+      
+      
+      
+      result: run_func = run_func("A name", 42, True)
+    `;
+    expect(result).toRenderTo(expected);
+  });
+
+  it("Method call without a reference and with call statement vars", () => {
+    const result = toSourceText(
+      <py.StatementList>
+        <py.FunctionCallExpression
+          target={"example_method"}
+          args={[
+            <py.VariableDeclaration name="name" initializer={"A name"} callStatementVar />,
+            <py.VariableDeclaration name="number" initializer={42} callStatementVar />,
+            <py.VariableDeclaration name="flag" initializer={true} callStatementVar />,
+          ]}
+        />
+      </py.StatementList>,
+    );
+    const expected = d`
+      example_method(name="A name", number=42, flag=True)
+    `;
+    expect(result).toRenderTo(expected);
+  });
+
+  it("Method call without a reference mixing unnamed and named vars", () => {
+    const result = toSourceText(
+      <py.StatementList>
+        <py.FunctionCallExpression
+          target={"example_method"}
+          args={[
+            <py.Value jsValue={"A name"} />,
+            <py.VariableDeclaration name="number" initializer={42} callStatementVar />,
+            <py.VariableDeclaration name="flag" initializer={true} callStatementVar />,
+          ]}
+        />
+      </py.StatementList>,
+    );
+    const expected = d`
+      example_method("A name", number=42, flag=True)
+    `;
+    expect(result).toRenderTo(expected);
   });
 });
